+++ brpc-java-core/src/main/java/com/baidu/brpc/server/push/ClientManager.java
@@ -1,0 +1,15 @@
+package com.baidu.brpc.server.push;
+
+/**
+ * 用于serverpush 注册client
+ */
+public interface ClientManager {
+
+    /**
+     * 注册和保存clientName
+     *
+     * @param clientName
+     */
+    Response registerClient(String clientName);
+
+}
