+++ brpc-java-examples/brpc-java-core-examples/src/main/java/com/baidu/brpc/example/push/userservice/UserPushApi.java
@@ -1,4 +1,4 @@
-package com.baidu.brpc.example.standard.userservice;
+package com.baidu.brpc.example.push.userservice;
 
 public interface UserPushApi {
